hotel_reservations_graph/pyproject.toml
[tool.poetry]
name = "hotel-reservations-graph"
version = "0.1.0"
description = ""
authors = ["Pedro Sousa <pedronsousa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.0.0"
bdd-llm = {path = "../bdd_llm", develop = true}
langchain-core = "^0.1.18"
langchain = "^0.1.5"
langchain-openai = "^0.0.5"
langgraph = "^0.0.21"
python-dotenv = "^1.0.1"
behave = "^1.2.6"
pyhamcrest = "^2.1.0"
langchain-experimental = "^0.0.50"
langchainhub = "^0.1.14"
langchain-together = "^0.0.2.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
